adds query that retrieves lab tests by date range
The locations is embedded as a json file. Querying for the location is
inefficient because they are recursive query. To get the district, we need to
make another query to the locations collection to get one that matches
the parentId of the current location we have. This means that we add 2 queries
for every lab test. If we have 50 lab tests, this is 100 extra queries. This is
painfully slow. The workaround we have is to embed the locations as a json
file as an inverted map. This makes retrieving the location by ID a constant
time operation.